fix: fix bug with answer in sprint game
