+++ src/pages/Games/Sprint/Question.tsx
@@ -1,25 +1,15 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { ROUND_TIME } from '../../../constants';
 import useFetchWords from '../../../hooks/useFetchWords';
-import { setCurrentWord, setStatus, setWords } from '../../../slices/sprint';
+import { setCurrentWord, setWords } from '../../../slices/sprint';
 import { TStore } from '../../../store';
 import Guess from './Guess';
 import Rightindicator from './Rightindicator';
 import { SprintGamePlay } from './styles';
+import Timer from './Timer';
 
 export default function Question() {
-  const [timer, setTimer] = useState(ROUND_TIME);
-  const [idInt, setidInt] = useState<NodeJS.Timer | null>(null);
-
   const dispatch = useDispatch();
-
-  const createTimer = () => {
-    const id = setInterval(() => {
-      setTimer((prevstate) => prevstate - 1);
-    }, 1000);
-    return id;
-  };
 
   const {
     group,
@@ -36,23 +26,11 @@ export default function Question() {
       dispatch(setCurrentWord({ word, id }));
     }
   }, [words, isSuccess]);
-  //timer logic
-  useEffect(() => {
-    if (!idInt) {
-      const id = createTimer();
-      setidInt(id);
-    }
-    if (timer === 0 && idInt) {
-      clearInterval(idInt);
-      setidInt(null);
-      dispatch(setStatus('ended'));
-    }
-  }, [timer]);
 
   return (
     <SprintGamePlay>
       <div className="question-wrapper">
-        <p className="time">{timer}</p>
+        <Timer />
         {wordsFromState.length > 0 && (
           <>
             <Rightindicator
